separate shape tests.
